+++ pkg/applier/cmd/exec_test.go
@@ -1,5 +1,4 @@
 // Copyright Contributors to the Open Cluster Management project
-
 package cmd
 
 import (
