+++ sonar-project.properties
@@ -1,7 +1,7 @@
 sonar.projectKey=open-cluster-management_applier
 sonar.projectName=applier
 sonar.sources=.
-sonar.exclusions=**/*_test.go,**/*_generated*.go,**/*_generated/**,**/vendor/**,./test/**
+sonar.exclusions=**/*_test.go,**/*_generated*.go,**/*_generated/**,**/vendor/**,./test/**,**/cmd.go
 sonar.tests=.
 sonar.test.inclusions=**/*_test.go
 sonar.test.exclusions=**/*_generated*.go,**/*_generated/**,**/vendor/**,./test/**
